fastapi/requirements.txt
aiosqlite==0.17.0
anyio==3.5.0
asgiref==3.5.0
asyncpg==0.25.0
atlastk==0.13
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
fastapi==0.75.1
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
iniconfig==1.1.1
iso8601==1.0.2
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pydantic==1.9.0
pyparsing==3.0.8
pypika-tortoise==0.1.4
pytest==7.1.1
pytest-asyncio==0.18.3
pytz==2022.1
PyYAML==6.0
rfc3986==1.5.0
sniffio==1.2.0
starlette==0.17.1
tomli==2.0.1
tortoise-orm==0.19.0
typing-extensions==4.1.1
uvicorn==0.17.6
